Support GroupsAccumulator accumulator for udaf
