+++ datafusion/physical-expr/src/aggregate/groups_accumulator/accumulate.rs
@@ -23,8 +23,7 @@ use arrow::datatypes::ArrowPrimitiveType;
 use arrow_array::{Array, BooleanArray, PrimitiveArray};
 use arrow_buffer::{BooleanBuffer, BooleanBufferBuilder, NullBuffer};
 
-use crate::EmitTo;
-
+use datafusion_expr::EmitTo;
 /// Track the accumulator null state per row: if any values for that
 /// group were null and if any values have been seen at all for that group.
 ///
